requirements.txt
coverage==5.5
Django==3.1.7
djangorestframework==3.12.2
psycopg2==2.8.6
django-nose==1.4.7
